+++ src/Engine/Globals.h
@@ -3,7 +3,7 @@
 
 namespace Global
 {
-	const std::string ENGINE_VERSION = "Alpha 0.055";
+	const std::string ENGINE_VERSION = "Alpha 0.06";
 };
 
 ///TODO: Add GUI support