+++ src/SpaceLevel.cpp
@@ -1,5 +1,7 @@
 #include "SpaceLevel.h"
 
+
+
 #include "Engine/Engine.h"
 #include "Engine/Input.h"
 
@@ -14,19 +16,14 @@ void SpaceLevel::Initialize(RendererBase* r, SoundManager* soundcontroller)
 	{
 		Spawn("TestEntity", new TestEntity(), Vector2(i, 200));
 	}
-
-
-
-
 }
 
 void SpaceLevel::Render()
 {
-	SDL_Color c = {0, 0, 0};
 	Renderer->RenderImage("Background", Vector2(0, 0));
+	SDL_Color c = {0, 0, 0};
 	Renderer->RenderDrawRect(Vector2(0, 0), Vector2(Renderer->TexSizeRepository["Background"].x,
 													Renderer->TexSizeRepository["Background"].y), c);
-
 	Super::Render();
 }