+++ src/TestEntity.cpp
@@ -11,79 +11,66 @@ void TestEntity::Render()
 	Renderer->RenderFont("I am a box", "ArialSmall", position.ToInteger()+Vector2(0, (int)dimensions.y), c);
 }
 
-void TestEntity::Tick()
+void TestEntity::Tick(float DeltaTime)
 {
-	velocity = velocity * 0.8;
+	velocity = velocity * 0.8 * DeltaTime;
 
 	Vector2 BackgroundDimensions = Renderer->TexSizeRepository["Background"];
 	if(position.x <= 0)
 	{
 		Delete();
-		velocity.x *= -1.0f;
+		velocity.x *= -1.0f * DeltaTime;
 		if(velocity.x < 0)
 			velocity.x = 1.5f;
 	}
 	else if(position.x >= BackgroundDimensions.x - dimensions.x)
 	{
-		velocity.x *= -1.0f;
+		velocity.x *= -1.0f * DeltaTime;
 		if(velocity.x > 0)
 			velocity.x = -1.5f;
 	}
 	if(position.y <= 0)
 	{
-		velocity.y *= -1.0f;
+		velocity.y *= -1.0f * DeltaTime;
 		if(velocity.y < 0)
 			velocity.y = 1.5f;
 	}
 	else if(position.y >= BackgroundDimensions.y - dimensions.y)
 	{
-		velocity.y *= -1.0f;
+		velocity.y *= -1.0f * DeltaTime;
 		if(velocity.y > 0)
 			velocity.y = -1.5f;
+	}
+
+	if(Input::GetKey(SDL_SCANCODE_UP))
+	{
+		velocity.y -= 5 * DeltaTime;
+	}
+	else if(Input::GetKey(SDL_SCANCODE_DOWN))
+	{
+		velocity.y += 5 * DeltaTime;
+	}
+
+	if(Input::GetKey(SDL_SCANCODE_LEFT))
+	{
+		velocity.x -= 5 * DeltaTime;
+	}
+	else if(Input::GetKey(SDL_SCANCODE_RIGHT))
+	{
+		velocity.x += 5 * DeltaTime;
 	}
 
 	position.x += velocity.x;
 	position.y += velocity.y;
 }
 
+// TODO: Get rid of Input() perhaps? (Maybe make it do the same thing as Tick with deltatime).
 void TestEntity::Input()
 {
-	if(Input::GetKey(SDL_SCANCODE_UP))
-	{
-		velocity.y -= 5;
-	}
-	else if(Input::GetKey(SDL_SCANCODE_DOWN))
-	{
-		velocity.y += 5;
-	}
-
-	if(Input::GetKey(SDL_SCANCODE_LEFT))
-	{
-		velocity.x -= 5;
-	}
-	else if(Input::GetKey(SDL_SCANCODE_RIGHT))
-	{
-		velocity.x += 5;
-	}
+	Super::Input();
 }
 
 void TestEntity::DefaultProperties()
 {
-	std::cout << "Linked list test from: TestEntity::DefaultProperties" << std::endl;
-
-	LinkedList<int> LinkyLink = LinkedList<int>();
-	LinkyLink.push_back(1);
-	LinkyLink.push_back(100);
-	LinkyLink.push_back(1000);
-	LinkyLink.push_back(10000);
-	LinkyLink.push_front(0);
-	LinkedNode<int>* A = LinkyLink.push_back(5);
-	LinkyLink.set_head(A);
-
-	LinkyLink.iterate([&] (LinkedNode<int>* node, int value)
-	{
-		std::cout << value << std::endl;
-	});
-
 	dimensions = Vector2(100, 100);
 }